refactor: chat
