+++ app/models/__init__.py
@@ -1,2 +1,2 @@
-from .history_chat import Chat
+from .chat import Chat
 from .sentiment_output import TodaySentimentReportOutput
